updated: w3provider to client only boundary and relevant fixes
